requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
opencv-python>=4.8.0
numpy>=1.26.0
scikit-learn>=1.3.0
joblib>=1.3.0
Werkzeug>=3.0.0
tensorflow-cpu>=2.15.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
pillow>=10.0.0 